refactor: move algolia routine to a helper file
